requirements.txt
agno
openai
pandas
python-dotenv